+++ MacPacker/Model/Archive2.swift
@@ -480,7 +480,12 @@ class Archive2: ObservableObject {
             if fileItemType == .file && isSupportedArchive(ext: url.pathExtension) {
                 fileItemType = .archive
             }
-            let fileItem = ArchiveItem(path: url, type: fileItemType, size: fileSize)
+            let fileItem = ArchiveItem(
+                path: url,
+                type: fileItemType,
+                compressedSize: fileSize,
+                uncompressedSize: fileSize
+            )
             
             if fileItemType == .directory {
                 resultDirectories.append(fileItem)