+++ MacPacker/Views/Preferences/GeneralSettingsView.swift
@@ -16,30 +16,47 @@ enum BreadcrumbPosition: String, CaseIterable {
 
 struct GeneralSettingsView: View {
     private let applicationSupportDirectory = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first
-    @AppStorage("setting.breadcrumbPosition") var breadcrumbPosition: BreadcrumbPosition = .bottom
+    @AppStorage(Keys.settingBreadcrumbPosition) var breadcrumbPosition: BreadcrumbPosition = .bottom
+    @AppStorage(Keys.showColumnCompressedSize) var showCompressedSize: Bool = true
+    @AppStorage(Keys.showColumnUncompressedSize) var showUncompressedSize: Bool = true
+    @AppStorage(Keys.showColumnModificationDate) var showModificationDate: Bool = true
+    @AppStorage(Keys.showColumnPosixPermissions) var showPermissions: Bool = false
     
     var body: some View {
-        VStack {
-            VStack {
-                VStack {
-                    HStack(alignment: .top) {
-                        Text("Breadcrumb position:")
-                            .frame(width: 160, alignment: .trailing)
-                        
-                        HStack {
-                            Picker("", selection: $breadcrumbPosition) {
-                                ForEach(BreadcrumbPosition.allCases, id: \.self) { position in
-                                    Text(position.rawValue)
-                                        .tag(position)
-                                }
-                            }
+        VStack(spacing: 8) {
+            HStack(alignment: .top) {
+                Text("Columns:")
+                    .frame(width: 160, alignment: .trailing)
+                
+                VStack(alignment: .leading) {
+                    Toggle("Packed Size", isOn: $showCompressedSize)
+                    Toggle("Size", isOn: $showUncompressedSize)
+                    Toggle("Modified", isOn: $showModificationDate)
+                    Toggle("Permissions", isOn: $showPermissions)
+                }
+                .padding(.leading, 8)
+                .toggleStyle(.checkbox)
+                .frame(width: 240, alignment: .leading)
+            }
+            
+            Divider()
+            
+            HStack(alignment: .top) {
+                Text("Breadcrumb position:")
+                    .frame(width: 160, alignment: .trailing)
+                
+                HStack {
+                    Picker("", selection: $breadcrumbPosition) {
+                        ForEach(BreadcrumbPosition.allCases, id: \.self) { position in
+                            Text(position.rawValue)
+                                .tag(position)
                         }
-                        .frame(width: 240, alignment: .leading)
                     }
                 }
+                .frame(width: 240, alignment: .leading)
             }
-            .padding()
         }
+        .padding()
     }
 }